Optional Inputs Argument
Adds a new inputs argument to the latex_document build rule.

This argument is to facilitate relative paths for files in latex files.
It will allow simulating building in a subdirectory of the repository
which matches the use case of most developers.

Allows developers using Bazel and those not using Bazel to share a
document without having to deal with the requirement of absolute paths
for the Bazel developer.

Specifying a list of directories in the inputs argument will add those directories
to the texinputs.